Replace 0 with "0" for sleep time in readme
If you copy/paste the example yaml files as they appear in README.md
you're likely to hit this error:

  mutation failed: cannot decode incoming new object: json: cannot
  unmarshal number into Go struct field ParamSpec.spec.params.default
  of type []string

I figure it's because the 0 is being treated as an integer not a
string.

It appears this was fixed long ago in the demo/04-tasks.yml, but
let's fix it here too for people using copy paste like I was.