Fix inconsistencies between demo yaml and readme
Adds the 'workingDir' key in a few places, and replaces '~' with
'/tekton/home' in one place.

While creating the previous commit I noticed some other
inconsistencies. For these changes I'm not sure what the impact is,
but I expect it's better to keep the readme and the demo files in
sync.

I used vimdiff to compare README.md with each demo file one at a
time, which worked reasonably well to highlight the differences.

Everything else matches pretty well, with one exception: The version
of demo/08-realworld.yml in the readme doesn't include the pipeline
definition, which I assumed might be intentional.